Made new lists for files (#5797)
* Made new lists for files

* New list order in design

* New texts
