+++ src/pages/registration/files_and_license_tab/FilesTableRow.tsx
@@ -10,10 +10,10 @@ import {
   FormControlLabel,
   FormHelperText,
   IconButton,
+  Link as MuiLink,
   ListItemIcon,
   ListItemText,
   MenuItem,
-  Link as MuiLink,
   Paper,
   Popover,
   Radio,
@@ -35,7 +35,7 @@ import { TruncatableTypography } from '../../../components/TruncatableTypography
 import { RootState } from '../../../redux/store';
 import { AssociatedFile, AssociatedFileType, FileRrs, FileVersion } from '../../../types/associatedArtifact.types';
 import { CustomerRrsType } from '../../../types/customerInstitution.types';
-import { LicenseUri, licenses } from '../../../types/license.types';
+import { licenses, LicenseUri } from '../../../types/license.types';
 import { SpecificFileFieldNames } from '../../../types/publicationFieldNames';
 import { Registration } from '../../../types/registration.types';
 import { dataTestId } from '../../../utils/dataTestIds';
@@ -67,6 +67,8 @@ export const FilesTableRow = ({ file, removeFile, baseFieldName, showFileVersion
   const embargoFieldName = `${baseFieldName}.${SpecificFileFieldNames.EmbargoDate}`;
   const legalNoteFieldName = `${baseFieldName}.${SpecificFileFieldNames.LegalNote}`;
   const rrsFieldName = `${baseFieldName}.${SpecificFileFieldNames.RightsRetentionStrategy}`;
+
+  const isArchived = file.type === 'UnpublishableFile';
 
   const isAcceptedFile = file.publisherVersion === FileVersion.Accepted;
   const rrsStrategy = file.rightsRetentionStrategy.configuredType ?? customer?.rightsRetentionStrategy.type;
@@ -100,8 +102,8 @@ export const FilesTableRow = ({ file, removeFile, baseFieldName, showFileVersion
       <TableRow
         data-testid={dataTestId.registrationWizard.files.fileRow}
         title={disabled ? t('registration.files_and_license.disabled_helper_text') : ''}
-        sx={{ bgcolor: disabled ? 'grey.400' : '', td: { pb: 0, borderBottom: 'unset' } }}>
-        <TableCell sx={{ minWidth: '13rem' }}>
+        sx={{ bgcolor: disabled ? 'grey.400' : '', td: !isArchived ? { pb: 0, borderBottom: 'unset' } : '' }}>
+        <TableCell sx={{ minWidth: '13rem', maxWidth: isArchived ? '20rem' : '' }}>
           <TruncatableTypography>{file.name}</TruncatableTypography>
         </TableCell>
 
@@ -133,7 +135,7 @@ export const FilesTableRow = ({ file, removeFile, baseFieldName, showFileVersion
           </Box>
         </TableCell>
 
-        <TableCell>{prettyBytes(file.size)}</TableCell>
+        <TableCell sx={{ minWidth: isArchived ? '5.5rem' : '' }}>{prettyBytes(file.size)}</TableCell>
 
         <TableCell>
           <Field name={administrativeAgreementFieldName}>
@@ -163,7 +165,7 @@ export const FilesTableRow = ({ file, removeFile, baseFieldName, showFileVersion
           </Field>
         </TableCell>
 
-        {showFileVersion && (
+        {showFileVersion && !isArchived && (
           <TableCell>
             <Field name={publisherVersionFieldName}>
               {({ field, meta: { error, touched } }: FieldProps<FileVersion | null>) => (
@@ -212,251 +214,254 @@ export const FilesTableRow = ({ file, removeFile, baseFieldName, showFileVersion
             </Field>
           </TableCell>
         )}
-
-        <TableCell>
-          <Field name={licenseFieldName}>
-            {({ field, meta: { error, touched } }: FieldProps<string>) => (
-              <TextField
-                id={field.name}
-                data-testid={dataTestId.registrationWizard.files.selectLicenseField}
-                sx={{ minWidth: '15rem' }}
-                select
-                disabled={disabled}
-                SelectProps={{
-                  renderValue: (option) => {
-                    const selectedLicense = licenses.find((license) => equalUris(license.id, option as string));
-                    return selectedLicense ? (
-                      <Box sx={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
-                        <img style={{ width: '5rem' }} src={selectedLicense.logo} alt={selectedLicense.name} />
-                        <span>{selectedLicense.name}</span>
-                      </Box>
-                    ) : null;
-                  },
-                }}
-                variant="filled"
-                value={licenses.find((license) => equalUris(license.id, field.value))?.id ?? ''}
-                error={!!error && touched}
-                helperText={<ErrorMessage name={field.name} />}
-                label={t('registration.files_and_license.conditions_for_using_file')}
-                required
-                onChange={({ target: { value } }) => setFieldValue(field.name, value)}>
-                {activeLicenses.map((license) => (
-                  <MenuItem
-                    data-testid={dataTestId.registrationWizard.files.licenseItem}
-                    key={license.id}
-                    value={license.id}
-                    divider
-                    dense
-                    sx={{ gap: '1rem' }}>
-                    <ListItemIcon>
-                      <img style={{ width: '5rem' }} src={license.logo} alt={license.name} />
-                    </ListItemIcon>
-                    <ListItemText>
-                      <Typography>{license.name}</Typography>
-                    </ListItemText>
-                  </MenuItem>
-                ))}
-                {!inactiveLicensesOpen && (
-                  <MenuItem
-                    data-testid={dataTestId.registrationWizard.files.licenseItemShowOlderVersion}
-                    sx={{ display: 'flex', justifyContent: 'center' }}
-                    onClickCapture={(e) => {
-                      e.stopPropagation();
-                      setInactiveLicensesOpen(!inactiveLicensesOpen);
-                    }}>
-                    <Typography sx={{ fontStyle: 'italic' }}>
-                      {t('registration.files_and_license.show_all_older_versions')}
-                    </Typography>
-                  </MenuItem>
-                )}
-                {inactiveLicenses.map((license) => (
-                  <MenuItem
-                    data-testid={dataTestId.registrationWizard.files.licenseItem}
-                    key={license.id}
-                    value={license.id}
-                    divider
-                    dense
-                    sx={{ gap: '1rem', display: inactiveLicensesOpen ? 'flex' : 'none' }}>
-                    <ListItemIcon>
-                      <img style={{ width: '5rem' }} src={license.logo} alt={license.name} />
-                    </ListItemIcon>
-                    <ListItemText>
-                      <Typography>{license.name}</Typography>
-                    </ListItemText>
-                  </MenuItem>
-                ))}
-              </TextField>
+        {!isArchived && (
+          <TableCell>
+            <Field name={licenseFieldName}>
+              {({ field, meta: { error, touched } }: FieldProps<string>) => (
+                <TextField
+                  id={field.name}
+                  data-testid={dataTestId.registrationWizard.files.selectLicenseField}
+                  sx={{ minWidth: '15rem' }}
+                  select
+                  disabled={disabled}
+                  SelectProps={{
+                    renderValue: (option) => {
+                      const selectedLicense = licenses.find((license) => equalUris(license.id, option as string));
+                      return selectedLicense ? (
+                        <Box sx={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
+                          <img style={{ width: '5rem' }} src={selectedLicense.logo} alt={selectedLicense.name} />
+                          <span>{selectedLicense.name}</span>
+                        </Box>
+                      ) : null;
+                    },
+                  }}
+                  variant="filled"
+                  value={licenses.find((license) => equalUris(license.id, field.value))?.id ?? ''}
+                  error={!!error && touched}
+                  helperText={<ErrorMessage name={field.name} />}
+                  label={t('registration.files_and_license.conditions_for_using_file')}
+                  required
+                  onChange={({ target: { value } }) => setFieldValue(field.name, value)}>
+                  {activeLicenses.map((license) => (
+                    <MenuItem
+                      data-testid={dataTestId.registrationWizard.files.licenseItem}
+                      key={license.id}
+                      value={license.id}
+                      divider
+                      dense
+                      sx={{ gap: '1rem' }}>
+                      <ListItemIcon>
+                        <img style={{ width: '5rem' }} src={license.logo} alt={license.name} />
+                      </ListItemIcon>
+                      <ListItemText>
+                        <Typography>{license.name}</Typography>
+                      </ListItemText>
+                    </MenuItem>
+                  ))}
+                  {!inactiveLicensesOpen && (
+                    <MenuItem
+                      data-testid={dataTestId.registrationWizard.files.licenseItemShowOlderVersion}
+                      sx={{ display: 'flex', justifyContent: 'center' }}
+                      onClickCapture={(e) => {
+                        e.stopPropagation();
+                        setInactiveLicensesOpen(!inactiveLicensesOpen);
+                      }}>
+                      <Typography sx={{ fontStyle: 'italic' }}>
+                        {t('registration.files_and_license.show_all_older_versions')}
+                      </Typography>
+                    </MenuItem>
+                  )}
+                  {inactiveLicenses.map((license) => (
+                    <MenuItem
+                      data-testid={dataTestId.registrationWizard.files.licenseItem}
+                      key={license.id}
+                      value={license.id}
+                      divider
+                      dense
+                      sx={{ gap: '1rem', display: inactiveLicensesOpen ? 'flex' : 'none' }}>
+                      <ListItemIcon>
+                        <img style={{ width: '5rem' }} src={license.logo} alt={license.name} />
+                      </ListItemIcon>
+                      <ListItemText>
+                        <Typography>{license.name}</Typography>
+                      </ListItemText>
+                    </MenuItem>
+                  ))}
+                </TextField>
+              )}
+            </Field>
+            {fileHasCustomerRrs && (
+              <Typography>
+                <Trans t={t} i18nKey="registration.files_and_license.institution_prefers_cc_by">
+                  {rrsPolicyLink}
+                </Trans>
+              </Typography>
             )}
-          </Field>
-          {fileHasCustomerRrs && (
-            <Typography>
-              <Trans t={t} i18nKey="registration.files_and_license.institution_prefers_cc_by">
-                {rrsPolicyLink}
-              </Trans>
-            </Typography>
-          )}
-          {fileHasOverriddenRrs && (
-            <Typography>
-              <Trans t={t} i18nKey="registration.files_and_license.opted_out_of_rrs">
-                {rrsPolicyLink}
-              </Trans>
-            </Typography>
-          )}
-        </TableCell>
+            {fileHasOverriddenRrs && (
+              <Typography>
+                <Trans t={t} i18nKey="registration.files_and_license.opted_out_of_rrs">
+                  {rrsPolicyLink}
+                </Trans>
+              </Typography>
+            )}
+          </TableCell>
+        )}
       </TableRow>
-      <TableRow
-        sx={{ bgcolor: disabled ? 'grey.400' : '' }}
-        title={disabled ? t('registration.files_and_license.disabled_helper_text') : ''}>
-        <TableCell sx={{ pt: 0, pb: 0 }} colSpan={showFileVersion ? 6 : 5}>
-          <Collapse in={openCollapsable}>
-            <Box
-              sx={{
-                m: '1rem 1rem 0 1rem',
-                display: 'grid',
-                gridTemplateColumns: '1fr auto',
-                gap: '2rem',
-              }}>
-              <Box sx={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}>
-                {isAcceptedFile && isNullRrs && (
-                  <FormControlLabel
-                    label={t('registration.files_and_license.mark_if_funder_requires_rrs')}
-                    control={
-                      <Checkbox
-                        checked={fileHasFunderRrs}
-                        onChange={() => {
-                          if (fileHasFunderRrs) {
-                            const nullRrsValue: FileRrs = {
-                              type: 'NullRightsRetentionStrategy',
-                              configuredType: rrsStrategy,
-                            };
-                            setFieldValue(rrsFieldName, nullRrsValue);
-                            setFieldValue(licenseFieldName, null);
-                          } else {
-                            const newRrsValue: FileRrs = {
-                              type: 'FunderRightsRetentionStrategy',
-                              configuredType: rrsStrategy,
-                            };
-                            setFieldValue(rrsFieldName, newRrsValue);
-                            setFieldValue(licenseFieldName, LicenseUri.CC_BY_4);
-                          }
-                        }}
-                      />
-                    }
-                  />
-                )}
+      {!isArchived && (
+        <TableRow
+          sx={{ bgcolor: disabled ? 'grey.400' : '' }}
+          title={disabled ? t('registration.files_and_license.disabled_helper_text') : ''}>
+          <TableCell sx={{ pt: 0, pb: 0 }} colSpan={showFileVersion ? 6 : 5}>
+            <Collapse in={openCollapsable}>
+              <Box
+                sx={{
+                  m: '1rem 1rem 0 1rem',
+                  display: 'grid',
+                  gridTemplateColumns: '1fr auto',
+                  gap: '2rem',
+                }}>
+                <Box sx={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}>
+                  {isAcceptedFile && isNullRrs && (
+                    <FormControlLabel
+                      label={t('registration.files_and_license.mark_if_funder_requires_rrs')}
+                      control={
+                        <Checkbox
+                          checked={fileHasFunderRrs}
+                          onChange={() => {
+                            if (fileHasFunderRrs) {
+                              const nullRrsValue: FileRrs = {
+                                type: 'NullRightsRetentionStrategy',
+                                configuredType: rrsStrategy,
+                              };
+                              setFieldValue(rrsFieldName, nullRrsValue);
+                              setFieldValue(licenseFieldName, null);
+                            } else {
+                              const newRrsValue: FileRrs = {
+                                type: 'FunderRightsRetentionStrategy',
+                                configuredType: rrsStrategy,
+                              };
+                              setFieldValue(rrsFieldName, newRrsValue);
+                              setFieldValue(licenseFieldName, LicenseUri.CC_BY_4);
+                            }
+                          }}
+                        />
+                      }
+                    />
+                  )}
 
-                {fileHasCustomerRrs && isCustomerRrs && (
-                  <Typography>
-                    {t('registration.files_and_license.institution_rights_policy_opt_out_instructions')}
-                  </Typography>
-                )}
+                  {fileHasCustomerRrs && isCustomerRrs && (
+                    <Typography>
+                      {t('registration.files_and_license.institution_rights_policy_opt_out_instructions')}
+                    </Typography>
+                  )}
 
-                {canOverrideRrs && (
-                  <FormControlLabel
-                    disabled={disabled}
-                    label={
-                      <Trans t={t} i18nKey="registration.files_and_license.follow_institution_rights_policy">
-                        {rrsPolicyLink}
-                      </Trans>
-                    }
-                    control={
-                      <Checkbox
-                        checked={!fileHasOverriddenRrs}
-                        onChange={() => {
-                          if (fileHasOverriddenRrs) {
-                            const customerRrsValue: FileRrs = {
-                              type: 'CustomerRightsRetentionStrategy',
-                              configuredType: rrsStrategy,
-                            };
-                            setFieldValue(rrsFieldName, customerRrsValue);
-                            setFieldValue(licenseFieldName, LicenseUri.CC_BY_4);
-                          } else {
-                            const overriddenRrsValue: FileRrs = {
-                              type: 'OverriddenRightsRetentionStrategy',
-                              configuredType: rrsStrategy,
-                            };
-                            setFieldValue(rrsFieldName, overriddenRrsValue);
-                            setFieldValue(licenseFieldName, null);
-                          }
-                        }}
-                      />
-                    }
-                  />
-                )}
+                  {canOverrideRrs && (
+                    <FormControlLabel
+                      disabled={disabled}
+                      label={
+                        <Trans t={t} i18nKey="registration.files_and_license.follow_institution_rights_policy">
+                          {rrsPolicyLink}
+                        </Trans>
+                      }
+                      control={
+                        <Checkbox
+                          checked={!fileHasOverriddenRrs}
+                          onChange={() => {
+                            if (fileHasOverriddenRrs) {
+                              const customerRrsValue: FileRrs = {
+                                type: 'CustomerRightsRetentionStrategy',
+                                configuredType: rrsStrategy,
+                              };
+                              setFieldValue(rrsFieldName, customerRrsValue);
+                              setFieldValue(licenseFieldName, LicenseUri.CC_BY_4);
+                            } else {
+                              const overriddenRrsValue: FileRrs = {
+                                type: 'OverriddenRightsRetentionStrategy',
+                                configuredType: rrsStrategy,
+                              };
+                              setFieldValue(rrsFieldName, overriddenRrsValue);
+                              setFieldValue(licenseFieldName, null);
+                            }
+                          }}
+                        />
+                      }
+                    />
+                  )}
 
-                {user?.isPublishingCurator && (
-                  <Field name={legalNoteFieldName}>
-                    {({ field }: FieldProps<string>) => (
-                      <TextField
+                  {user?.isPublishingCurator && (
+                    <Field name={legalNoteFieldName}>
+                      {({ field }: FieldProps<string>) => (
+                        <TextField
+                          {...field}
+                          value={field.value ?? ''}
+                          data-testid={dataTestId.registrationWizard.files.legalNoteField}
+                          variant="filled"
+                          fullWidth
+                          label={t('registration.files_and_license.legal_note')}
+                          multiline
+                          disabled={disabled}
+                          helperText={<ErrorMessage name={field.name} />}
+                        />
+                      )}
+                    </Field>
+                  )}
+                </Box>
+
+                <Box sx={{ display: 'flex', gap: '0.5rem', alignItems: 'center', alignSelf: 'end' }}>
+                  <Field name={embargoFieldName}>
+                    {({ field, meta: { error, touched } }: FieldProps) => (
+                      <DatePicker
                         {...field}
-                        value={field.value ?? ''}
-                        data-testid={dataTestId.registrationWizard.files.legalNoteField}
-                        variant="filled"
-                        fullWidth
-                        label={t('registration.files_and_license.legal_note')}
-                        multiline
-                        disabled={disabled}
-                        helperText={<ErrorMessage name={field.name} />}
+                        label={t('registration.files_and_license.embargo')}
+                        value={field.value ? new Date(field.value) : null}
+                        onChange={(date) => setFieldValue(field.name, date ?? '')}
+                        maxDate={new Date(new Date().getFullYear() + 5, 11, 31)}
+                        disabled={file.administrativeAgreement || disabled}
+                        sx={{ minWidth: '15rem' }}
+                        slotProps={{
+                          textField: {
+                            inputProps: { 'data-testid': dataTestId.registrationWizard.files.embargoDateField },
+                            variant: 'filled',
+                            onBlur: () => !touched && setFieldTouched(field.name),
+                            error: !!error && touched,
+                            helperText: <ErrorMessage name={field.name} />,
+                          },
+                        }}
                       />
                     )}
                   </Field>
-                )}
-              </Box>
 
-              <Box sx={{ display: 'flex', gap: '0.5rem', alignItems: 'center', alignSelf: 'end' }}>
-                <Field name={embargoFieldName}>
-                  {({ field, meta: { error, touched } }: FieldProps) => (
-                    <DatePicker
-                      {...field}
-                      label={t('registration.files_and_license.embargo')}
-                      value={field.value ? new Date(field.value) : null}
-                      onChange={(date) => setFieldValue(field.name, date ?? '')}
-                      maxDate={new Date(new Date().getFullYear() + 5, 11, 31)}
-                      disabled={file.administrativeAgreement || disabled}
-                      sx={{ minWidth: '15rem' }}
-                      slotProps={{
-                        textField: {
-                          inputProps: { 'data-testid': dataTestId.registrationWizard.files.embargoDateField },
-                          variant: 'filled',
-                          onBlur: () => !touched && setFieldTouched(field.name),
-                          error: !!error && touched,
-                          helperText: <ErrorMessage name={field.name} />,
-                        },
-                      }}
-                    />
-                  )}
-                </Field>
-
-                <Tooltip title={t('common.help')}>
-                  <IconButton onClick={(event) => setEmbargoPopperAnchorEl(event.currentTarget)}>
-                    <HelpOutlineIcon />
-                  </IconButton>
-                </Tooltip>
-                <Popover
-                  open={!!embargoPopperAnchorEl}
-                  anchorEl={embargoPopperAnchorEl}
-                  onClose={() => setEmbargoPopperAnchorEl(null)}
-                  anchorOrigin={{
-                    vertical: 'bottom',
-                    horizontal: 'left',
-                  }}>
-                  <Paper sx={{ p: '1rem' }}>
-                    <Typography>{t('registration.files_and_license.file_publish_date_helper_text')}</Typography>
-                  </Paper>
-                </Popover>
+                  <Tooltip title={t('common.help')}>
+                    <IconButton onClick={(event) => setEmbargoPopperAnchorEl(event.currentTarget)}>
+                      <HelpOutlineIcon />
+                    </IconButton>
+                  </Tooltip>
+                  <Popover
+                    open={!!embargoPopperAnchorEl}
+                    anchorEl={embargoPopperAnchorEl}
+                    onClose={() => setEmbargoPopperAnchorEl(null)}
+                    anchorOrigin={{
+                      vertical: 'bottom',
+                      horizontal: 'left',
+                    }}>
+                    <Paper sx={{ p: '1rem' }}>
+                      <Typography>{t('registration.files_and_license.file_publish_date_helper_text')}</Typography>
+                    </Paper>
+                  </Popover>
+                </Box>
               </Box>
+            </Collapse>
+            <Box sx={{ display: 'flex', justifyContent: 'space-around' }}>
+              <IconButton
+                onClick={() => setOpenCollapsable(!openCollapsable)}
+                size="small"
+                data-testid={dataTestId.registrationWizard.files.expandFileRowButton}>
+                {openCollapsable ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
+              </IconButton>
             </Box>
-          </Collapse>
-          <Box sx={{ display: 'flex', justifyContent: 'space-around' }}>
-            <IconButton
-              onClick={() => setOpenCollapsable(!openCollapsable)}
-              size="small"
-              data-testid={dataTestId.registrationWizard.files.expandFileRowButton}>
-              {openCollapsable ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-            </IconButton>
-          </Box>
-        </TableCell>
-      </TableRow>
+          </TableCell>
+        </TableRow>
+      )}
     </>
   );
 };
