+++ src/translations/nbTranslations.json
@@ -1286,6 +1286,7 @@
       "file_publish_date_helper_text": "Filen er låst fram til angitt dato",
       "files": "Filer",
       "files_are_not_published": "Filer publiseres ikke",
+      "files_to_publish": "Venter på publisering",
       "files_awaits_approval_one": "{{count}} fil og valgt lisens venter på å bli verifisert",
       "files_awaits_approval_other": "{{count}} filer og valgte lisenser venter på å bli verifisert",
       "files_awaits_approval_unknown": "Venter på verifisering av fil(er)",
@@ -1307,6 +1308,7 @@
       "open_file": "Åpne fil",
       "opted_out_of_rrs": "Du har valgt å ikke følge vedtatt <0>rettighetspolitikk</0>",
       "published": "Publisert",
+      "published_files": "Publiserte filer",
       "published_version": "Publisert versjon",
       "registration_with_file": "Resultat med fil",
       "registration_without_file": "Resultat uten fil",
